refactor: 💡 move a function from scoresummary to playerstatus
